requirements.txt
Django
django-cors-headers
djangorestframework
firebase-admin